WiP Struct with array attributes
