+++ spec/struct_spec.rb
@@ -2,89 +2,7 @@
 
 require 'spec_helper'
 require 'plumb'
-
-module Plumb
-  class Struct
-    attr_reader :errors, :attributes
-
-    def initialize(attrs = {})
-      @errors = {}
-      @attributes = self.class.attribute_specs.each_with_object({}) do |(name, type), hash|
-        name = name.to_sym
-        value = attrs.key?(name) ? attrs[name] : Plumb::Undefined
-        result = type.resolve(value)
-        @errors[name] = result.errors unless result.valid?
-        hash[name] = result.value
-      end
-    end
-
-    def valid? = errors.none?
-
-    def inspect
-      %(#<#{self.class}:#{object_id} #{attributes.map { |k, v| [k, v.inspect].join(': ') }.join(' ')}>)
-    end
-
-    class << self
-      def call(result)
-        return result.invalid(errors: ['Must be a Hash of attributes']) unless result.value.is_a?(Hash)
-
-        instance = new(result.value)
-        instance.valid? ? result.valid(instance) : result.invalid(instance, errors: instance.errors)
-      end
-
-      def attribute_specs
-        @attribute_specs ||= {}
-      end
-
-      # attribute(:friend) { attribute(:name, String) }
-      # attribute(:friend, MyStruct) { attribute(:name, String) }
-      # attribute(:name, String)
-      # attribute(:friends, Types::Array) { attribute(:name, String) }
-      # attribute(:friends, Types::Array)
-      #
-      def attribute(name, *args, &block)
-        name = name.to_sym
-
-        type = case args
-               in [Class => klass] if block_given? # attribute(:friend, MyStruct) { ... }
-                 sub = Class.new(klass)
-                 sub.instance_exec(&block)
-                 __set_nested_class__(name, sub)
-                 Plumb::Steppable.wrap(sub)
-               in [] if block_given? # attribute(:friend) { ... }
-                 sub = Class.new(Plumb::Struct)
-                 sub.instance_exec(&block)
-                 __set_nested_class__(name, sub)
-                 Plumb::Steppable.wrap(sub)
-               in [Plumb::ArrayClass => type] if block_given?
-                 sub = type.element_type
-                 if sub == Plumb::Types::Any
-                   sub = Class.new(Plumb::Struct)
-                   sub.instance_exec(&block)
-                   __set_nested_class__(name, sub)
-                   Plumb::Types::Array[Plumb::Steppable.wrap(sub)]
-                 else
-                   type
-                 end
-               in [Plumb::Steppable => type] # attribute(:name, String)
-                 type
-               in [type]
-                 Plumb::Steppable.wrap(type)
-               else
-                 raise ArgumentError, "Invalid arguments: #{args.inspect}"
-               end
-
-        attribute_specs[name] = type
-        define_method(name) { @attributes[name] }
-      end
-
-      def __set_nested_class__(name, klass)
-        name = name.to_s.split('_').map(&:capitalize).join.sub(/s$/, '')
-        const_set(name, klass) unless const_defined?(name)
-      end
-    end
-  end
-end
+require 'plumb/struct'
 
 module Types
   class User < Plumb::Struct
@@ -102,6 +20,17 @@ module Types
     attribute :books, Array do
       attribute :isbn, String
     end
+
+    def book_count = books.size
+  end
+
+  class StaffMember < Plumb::Struct
+    attribute :name, String
+    attribute :age, Lax::Integer[18..]
+  end
+
+  class Office < Plumb::Struct
+    attribute :staff, Array[StaffMember]
   end
 end
 
@@ -127,6 +56,22 @@ RSpec.describe Plumb::Struct do
     expect(user.friend.email).to eq 'john@server.com'
     expect(user.company.name).to eq 'Acme'
     expect(user.books.map(&:isbn)).to eq ['123']
+    expect(user.book_count).to eq 1
+  end
+
+  specify 'Array[StructType]' do
+    office = Types::Office.new(staff: [{ name: 'Jane', age: '20' }])
+    expect(office.valid?).to be true
+    expect(office.staff.first.name).to eq 'Jane'
+    expect(office.staff.first.age).to eq 20
+  end
+
+  specify 'Array[StructType] with errors' do
+    office = Types::Office.new(staff: [{ name: 'Jane' }])
+    expect(office.valid?).to be false
+    expect(office.staff.first.name).to eq 'Jane'
+    expect(office.staff.first.valid?).to be(false)
+    expect(office.staff.first.errors[:age]).not_to be_empty
   end
 
   specify 'invalid' do
